refactor: move MockMetricsRecorder to mock package and rename it
Move mock metrics recorder implementation to dedicated mock subpackage
for better code organization and testability.

- Relocate MockMetricsRecorder to pkg/scheduler/framework/runtime/mock and rename it to MetricRecorder
- Update test imports to reference new mock package location
- Improve ObserveInFlightEventsAsync documentation
