+++ pkg/scheduler/framework/runtime/framework_test.go
@@ -47,6 +47,7 @@ import (
 	internalqueue "k8s.io/kubernetes/pkg/scheduler/backend/queue"
 	"k8s.io/kubernetes/pkg/scheduler/framework"
 	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/names"
+	"k8s.io/kubernetes/pkg/scheduler/framework/runtime/mock"
 	"k8s.io/kubernetes/pkg/scheduler/metrics"
 	st "k8s.io/kubernetes/pkg/scheduler/testing"
 	"k8s.io/utils/ptr"
@@ -4251,7 +4252,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 
 				// Expected metrics fields
 				expectedPluginDurationCallCount int
-				expectedPluginDurationRecords   []PluginDurationRecord
+				expectedPluginDurationRecords   []mock.PluginDurationRecord
 
 				// Plugin configuration fields
 				prefilterPlugins *config.PluginSet
@@ -4270,7 +4271,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Success,
 					prefilterPlugins:                &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin, Weight: 1}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "PreFilter",
 							PluginName:     testPlugin,
@@ -4287,7 +4288,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Success,
 					scorePlugins:                    &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin, Weight: 1}}},
 					expectedPluginDurationCallCount: 2, // Score runs once per node (2 nodes)
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "Score",
 							PluginName:     testPlugin,
@@ -4307,7 +4308,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Success,
 					bindPlugins:                     &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "Bind",
 							PluginName:     testPlugin,
@@ -4323,7 +4324,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Error,
 					prefilterPlugins:                &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin, Weight: 1}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "PreFilter",
 							PluginName:     testPlugin,
@@ -4339,7 +4340,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Error,
 					bindPlugins:                     &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "Bind",
 							PluginName:     testPlugin,
@@ -4354,7 +4355,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Success,
 					prescorePlugins:                 &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin, Weight: 1}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "PreScore",
 							PluginName:     testPlugin,
@@ -4370,7 +4371,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Error,
 					prescorePlugins:                 &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin, Weight: 1}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "PreScore",
 							PluginName:     testPlugin,
@@ -4388,7 +4389,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Error,
 					scorePlugins:                    &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin, Weight: 1}}},
 					expectedPluginDurationCallCount: 2, // Score still runs for all nodes even on error
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "Score",
 							PluginName:     testPlugin,
@@ -4408,7 +4409,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Success,
 					reservePlugins:                  &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "Reserve",
 							PluginName:     testPlugin,
@@ -4424,7 +4425,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Error,
 					reservePlugins:                  &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "Reserve",
 							PluginName:     testPlugin,
@@ -4439,7 +4440,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Success,
 					reservePlugins:                  &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "Unreserve",
 							PluginName:     testPlugin,
@@ -4454,7 +4455,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Success,
 					prebindPlugins:                  &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "PreBind",
 							PluginName:     testPlugin,
@@ -4470,7 +4471,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Error,
 					prebindPlugins:                  &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "PreBind",
 							PluginName:     testPlugin,
@@ -4485,7 +4486,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Success,
 					postbindPlugins:                 &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "PostBind",
 							PluginName:     testPlugin,
@@ -4500,7 +4501,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Success,
 					permitPlugins:                   &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "Permit",
 							PluginName:     testPlugin,
@@ -4516,7 +4517,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Error,
 					permitPlugins:                   &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "Permit",
 							PluginName:     testPlugin,
@@ -4532,7 +4533,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 					wantStatus:                      fwk.Wait,
 					permitPlugins:                   &config.PluginSet{Enabled: []config.Plugin{{Name: testPlugin}}},
 					expectedPluginDurationCallCount: 1,
-					expectedPluginDurationRecords: []PluginDurationRecord{
+					expectedPluginDurationRecords: []mock.PluginDurationRecord{
 						{
 							ExtensionPoint: "Permit",
 							PluginName:     testPlugin,
@@ -4545,7 +4546,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 			for _, tt := range tests {
 				t.Run(tt.name, func(t *testing.T) {
 					// Create a fake metrics recorder to verify calls
-					mockRecorder := NewMockMetricsRecorder()
+					mockRecorder := mock.NewMetricsRecorder()
 
 					// Create registry with test plugin
 					plugin := &TestPlugin{name: testPlugin, inj: tt.inject}
@@ -4611,7 +4612,7 @@ func TestRecordingMetricsWithMocks(t *testing.T) {
 						// Note: Ignore the Value field as it contains timing information
 						if diff := cmp.Diff(tt.expectedPluginDurationRecords,
 							mockRecorder.GetPluginDurationRecords(),
-							cmpopts.IgnoreFields(PluginDurationRecord{}, "Value")); diff != "" {
+							cmpopts.IgnoreFields(mock.PluginDurationRecord{}, "Value")); diff != "" {
 							t.Errorf("plugin duration records mismatch (-want +got):\n%s", diff)
 						}
 					} else {
